third_party/blink/web_tests/external/wpt/content-security-policy/reporting-api/report-to-directive-allowed-in-meta.https.sub-expected.txt

This is a testharness.js-based test.
[FAIL] Report is observable to ReportingObserver
  assert_equals: expected (object) null but got (string) "https://web-platform.test:8444/content-security-policy/reporting-api/report-to-directive-allowed-in-meta.https.sub.html"
Harness: the test ran to completion.
